apps/iot/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "pms-iot"
version = "0.1.0"
description = "PMS IoT Service — device lifecycle, EMQX auth, ThingsBoard sync, Headscale SSH"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "motor>=3.6.0",
    "beanie>=1.27.0",
    "redis[asyncio]>=5.2.0",
    "aio-pika>=9.5.0",
    "aioboto3>=13.2.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",
    "httpx>=0.28.0",
    "aiomqtt>=2.3.0",
    "email-validator>=2.2.0",
    "asyncssh>=2.17.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "fakeredis>=2.26.0",
    "mongomock-motor>=0.0.34",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
